Refactor ExecutionPlanType
