+++ tornado-api/src/main/java/uk/ac/manchester/tornado/api/ExecutionPlanType.java
@@ -1,0 +1,80 @@
+/*
+ * Copyright (c) 2024, APT Group, Department of Computer Science,
+ * The University of Manchester.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+package uk.ac.manchester.tornado.api;
+
+import uk.ac.manchester.tornado.api.plan.types.OffConcurrentDevices;
+import uk.ac.manchester.tornado.api.plan.types.OffMemoryLimit;
+import uk.ac.manchester.tornado.api.plan.types.OffPrintKernel;
+import uk.ac.manchester.tornado.api.plan.types.OffProfiler;
+import uk.ac.manchester.tornado.api.plan.types.OffThreadInfo;
+import uk.ac.manchester.tornado.api.plan.types.WithBatch;
+import uk.ac.manchester.tornado.api.plan.types.WithClearProfiles;
+import uk.ac.manchester.tornado.api.plan.types.WithCompilerFlags;
+import uk.ac.manchester.tornado.api.plan.types.WithConcurrentDevices;
+import uk.ac.manchester.tornado.api.plan.types.WithDefaultScheduler;
+import uk.ac.manchester.tornado.api.plan.types.WithDevicePlan;
+import uk.ac.manchester.tornado.api.plan.types.WithDynamicReconfiguration;
+import uk.ac.manchester.tornado.api.plan.types.WithFreeDeviceMemory;
+import uk.ac.manchester.tornado.api.plan.types.WithGridScheduler;
+import uk.ac.manchester.tornado.api.plan.types.WithMemoryLimit;
+import uk.ac.manchester.tornado.api.plan.types.WithPrintKernel;
+import uk.ac.manchester.tornado.api.plan.types.WithProfiler;
+import uk.ac.manchester.tornado.api.plan.types.WithResetDevice;
+import uk.ac.manchester.tornado.api.plan.types.WithThreadInfo;
+import uk.ac.manchester.tornado.api.plan.types.WithWarmUp;
+
+public abstract sealed class ExecutionPlanType extends TornadoExecutionPlan //
+        permits OffConcurrentDevices, OffMemoryLimit, OffPrintKernel, //
+        OffProfiler, OffThreadInfo, WithWarmUp, WithBatch, WithClearProfiles,  //
+        WithCompilerFlags, WithConcurrentDevices, WithDefaultScheduler, //
+        WithDevicePlan, WithDynamicReconfiguration, WithFreeDeviceMemory, //
+        WithGridScheduler, WithMemoryLimit, WithPrintKernel, WithProfiler, //
+        WithResetDevice, WithThreadInfo {
+
+    public ExecutionPlanType(TornadoExecutionPlan parentNode) {
+
+        // Set link between the previous action (parent) and the new one
+        this.parentLink = parentNode;
+        updateChildFromRoot(this);
+
+        // Copy the reference for the executor
+        this.tornadoExecutor = parentNode.tornadoExecutor;
+
+        // Copy the reference for the execution frame
+        this.executionFrame = parentNode.executionFrame;
+
+        // Set child reference to this instance
+        this.childLink = this;
+    }
+
+    private void updateChildFromRoot(TornadoExecutionPlan childNode) {
+        assert childNode != null;
+        TornadoExecutionPlan rootNode = childNode;
+        TornadoExecutionPlan iterator = childNode;
+
+        // Traverse the list until we find the root node
+        while (iterator != null) {
+            rootNode = iterator;
+            iterator = iterator.parentLink;
+        }
+
+        // Set the child of the root node to the new node
+        rootNode.childLink = childNode;
+    }
+
+}
