+++ tornado-api/src/main/java/uk/ac/manchester/tornado/api/plan/types/WithThreadInfo.java
@@ -17,6 +17,7 @@
  */
 package uk.ac.manchester.tornado.api.plan.types;
 
+import uk.ac.manchester.tornado.api.ExecutionPlanType;
 import uk.ac.manchester.tornado.api.TornadoExecutionPlan;
 
 public final class WithThreadInfo extends ExecutionPlanType {
